+++ src/plugin/Actions.php
@@ -11,7 +11,7 @@ class Actions
 	 * @param array $links all links of the plugin
 	 * @return array links with added settings link
 	 */
-	public function scouting_oidc_actions_plugin_links($links) {
+	public function scouting_oidc_actions_plugin_links(array $links): array {
 		array_unshift($links, '<a href="'.esc_url(get_admin_url(null, 'admin.php?page=scouting-oidc-settings')).'">' . __("Settings", "scouting-openid-connect") . '</a>');
 		return $links;
 	}