+++ src/plugin/Description.php
@@ -11,7 +11,7 @@ class Description
 	 * @param array $all_plugins all plugins with their information
 	 * @return array All plugins with their information including the modified description
 	 */
-	public function scouting_oidc_description_modify_plugin($all_plugins) {
+	public function scouting_oidc_description_modify_plugin(array $all_plugins): array {
 		if (isset($all_plugins['scouting-openid-connect/scouting-openid-connect.php'])) {
 			$description = __('WordPress plugin for logging in with Scouting Nederland OpenID Connect Server.', 'scouting-openid-connect');
 			$all_plugins['scouting-openid-connect/scouting-openid-connect.php']['Description'] = $description;