+++ src/settings/General.php
@@ -5,7 +5,7 @@ if ( ! defined( 'ABSPATH' ) ) exit; // Exit if accessed directly
 
 class Settings_General
 {
-    public function scouting_oidc_settings_general() {
+    public function scouting_oidc_settings_general(): void {
         // Add settings sections
         add_settings_section(
             'scouting_oidc_general_settings',                   // ID
@@ -80,6 +80,15 @@ class Settings_General
             [$this, 'scouting_oidc_settings_general_user_auto_create_callback'], // Callback to render field
             'scouting-openid-connect-settings',                                  // Page slug
             'scouting_oidc_general_settings'                                     // Section ID where the field should be added
+        );
+
+        // Add a settings selectbox field
+        add_settings_field(
+            'scouting_oidc_user_duplicate_email',                               // Field ID
+            __('When an email already exists', 'scouting-openid-connect'),      // Field label
+            [$this, 'scouting_oidc_settings_general_duplicate_email_callback'], // Callback to render field
+            'scouting-openid-connect-settings',                                 // Page slug
+            'scouting_oidc_general_settings'                                    // Section ID where the field should be added
         );
 
         // Add a settings checkbox field
@@ -176,6 +185,15 @@ class Settings_General
                 'sanitize_callback' => [$this, 'scouting_oidc_sanitize_boolean_option'] // Sanitize the input value as a boolean (0 or 1)
             ]
         );
+
+        // Register settings
+        register_setting(
+            'scouting_oidc_settings_group',                                                     // Settings group name
+            'scouting_oidc_user_duplicate_email',                                               // Option name
+            [
+                'sanitize_callback' => [$this, 'scouting_oidc_sanitize_duplicate_email_option'] // Sanitize the input value for duplicate email handling
+            ]
+        );
         
         // Register settings
         register_setting(
@@ -206,7 +224,7 @@ class Settings_General
     }
 
     // Sanitize the display name option
-    public function scouting_oidc_sanitize_display_name_option($input) {
+    public function scouting_oidc_sanitize_display_name_option(mixed $input): string {
         // Define allowed options
         $valid = ['fullname', 'firstname', 'lastname'];
         
@@ -215,12 +233,21 @@ class Settings_General
     }
 
     // Sanitize the input value as boolean
-    public function scouting_oidc_sanitize_boolean_option($input) {
+    public function scouting_oidc_sanitize_boolean_option(mixed $input): int {
         return $input ? 1 : 0;
     }
 
+    // Sanitize the duplicate email option
+    public function scouting_oidc_sanitize_duplicate_email_option(mixed $input): string {
+        // Define allowed options
+        $valid = ['plus_addressing', 'error'];
+
+        // Return the input if it’s a valid option; otherwise, default to 'plus_addressing'
+        return in_array($input, $valid, true) ? $input : 'plus_addressing';
+    }
+
     // Sanitize the login redirect option
-    public function scouting_oidc_sanitize_login_redirect_option($input) {
+    public function scouting_oidc_sanitize_login_redirect_option(mixed $input): string {
         // Define allowed options
         $valid = ['default', 'frontpage', 'dashboard', 'custom'];
         
@@ -229,7 +256,7 @@ class Settings_General
     }
 
     // Sanitize the custom redirect option
-    public function scouting_oidc_sanitize_custom_redirect_option($input) {
+    public function scouting_oidc_sanitize_custom_redirect_option(mixed $input): string {
         // Define your fixed base domain
         $base_domain = home_url('/');
 
@@ -243,10 +270,10 @@ class Settings_General
     }
 
     // Callback to render section content
-    public function scouting_oidc_settings_general_callback() {}
+    public function scouting_oidc_settings_general_callback(): void {}
 
     // Callback to render selectbox field
-    public function scouting_oidc_settings_general_display_name_callback() {
+    public function scouting_oidc_settings_general_display_name_callback(): void {
         $possible_values = array(
             'fullname' => __('Full name', 'scouting-openid-connect'),
             'firstname' => __('First name', 'scouting-openid-connect'),
@@ -265,7 +292,7 @@ class Settings_General
     }
 
     // Callback to render checkbox field
-    public function scouting_oidc_settings_general_birthdate_callback() {
+    public function scouting_oidc_settings_general_birthdate_callback(): void {
         if (get_option('scouting_oidc_user_birthdate'))
             echo '<input type="checkbox" id="scouting_oidc_user_birthdate" name="scouting_oidc_user_birthdate" checked/>';
         else
@@ -273,7 +300,7 @@ class Settings_General
     }
 
     // Callback to render checkbox field
-    public function scouting_oidc_settings_general_gender_callback() {
+    public function scouting_oidc_settings_general_gender_callback(): void {
         if (get_option('scouting_oidc_user_gender'))
             echo '<input type="checkbox" id="scouting_oidc_user_gender" name="scouting_oidc_user_gender" checked/>';
         else 
@@ -281,7 +308,7 @@ class Settings_General
     }
 
     // Callback to render checkbox field
-    public function scouting_oidc_settings_general_phone_callback() {
+    public function scouting_oidc_settings_general_phone_callback(): void {
         if (get_option('scouting_oidc_user_phone'))
             echo '<input type="checkbox" id="scouting_oidc_user_phone" name="scouting_oidc_user_phone" checked/>';
         else 
@@ -289,7 +316,7 @@ class Settings_General
     }
 
     // Callback to render checkbox field
-    public function scouting_oidc_settings_general_address_callback() {
+    public function scouting_oidc_settings_general_address_callback(): void {
         if (get_option('scouting_oidc_user_address'))
             echo '<input type="checkbox" id="scouting_oidc_user_address" name="scouting_oidc_user_address" checked/>';
         else 
@@ -297,7 +324,7 @@ class Settings_General
     }
 
     // Callback to render checkbox field
-    public function scouting_oidc_settings_general_woocommerce_sync_callback() {
+    public function scouting_oidc_settings_general_woocommerce_sync_callback(): void {
         if (get_option('scouting_oidc_user_woocommerce_sync'))
             echo '<input type="checkbox" id="scouting_oidc_user_woocommerce_sync" name="scouting_oidc_user_woocommerce_sync" checked/>';
         else 
@@ -305,15 +332,33 @@ class Settings_General
     }
 
     // Callback to render checkbox field
-    public function scouting_oidc_settings_general_user_auto_create_callback() {
+    public function scouting_oidc_settings_general_user_auto_create_callback(): void {
         if (get_option('scouting_oidc_user_auto_create'))
             echo '<input type="checkbox" id="scouting_oidc_user_auto_create" name="scouting_oidc_user_auto_create" checked/>';
         else
             echo '<input type="checkbox" id="scouting_oidc_user_auto_create" name="scouting_oidc_user_auto_create"/>';
     }
 
+    // Callback to render selectbox field
+    public function scouting_oidc_settings_general_duplicate_email_callback(): void {
+        $possible_values = array(
+            'plus_addressing' => __('Add plus addressing to email', 'scouting-openid-connect'),
+            'error' => __('Stop creation of user with an error', 'scouting-openid-connect'),
+        );
+        $value = get_option('scouting_oidc_user_duplicate_email', 'plus_addressing');
+
+        echo '<select id="scouting_oidc_user_duplicate_email" name="scouting_oidc_user_duplicate_email" style="width: 310px;">';
+        foreach ($possible_values as $key => $name) {
+            if ($value == $key)
+                echo '<option value="' . esc_attr($key) . '" selected>' . esc_html($name) . '</option>';
+            else
+                echo '<option value="' . esc_attr($key) . '">' . esc_html($name) . '</option>';
+        }
+        echo '</select>';
+    }
+
     // Callback to render checkbox field
-    public function scouting_oidc_settings_general_user_redirect_callback() {
+    public function scouting_oidc_settings_general_user_redirect_callback(): void {
         if (get_option('scouting_oidc_user_redirect'))
             echo '<input type="checkbox" id="scouting_oidc_user_redirect" name="scouting_oidc_user_redirect" checked/>';
         else
@@ -321,7 +366,7 @@ class Settings_General
     }
 
     // Callback to render selectbox field
-    public function scouting_oidc_settings_general_login_redirect_callback() {
+    public function scouting_oidc_settings_general_login_redirect_callback(): void {
         $possible_values = array(
             'default' => __('Default (no action)', 'scouting-openid-connect'),
             'frontpage' => __('Frontpage', 'scouting-openid-connect'),
@@ -341,7 +386,7 @@ class Settings_General
     }
 
     // Callback to render text field
-    public function scouting_oidc_settings_general_custom_redirect_callback() {
+    public function scouting_oidc_settings_general_custom_redirect_callback(): void {
         $value = get_option('scouting_oidc_custom_redirect');
 
         // Define your fixed base domain