+++ src/settings/Page.php
@@ -3,15 +3,15 @@ namespace ScoutingOIDC;
 
 if ( ! defined( 'ABSPATH' ) ) exit; // Exit if accessed directly
 
-require_once plugin_dir_path(__FILE__) . 'oidc.php';
-require_once plugin_dir_path(__FILE__) . 'general.php';
+require_once plugin_dir_path(__FILE__) . 'Oidc.php';
+require_once plugin_dir_path(__FILE__) . 'General.php';
 
 use ScoutingOIDC\Settings_Oidc;
 use ScoutingOIDC\Settings_General;
 
 class Settings
 {
-    public function scouting_oidc_settings_submenu_page() {
+    public function scouting_oidc_settings_submenu_page(): void {
         add_submenu_page(
             'scouting-oidc-settings',                        // Parent slug (matches the main menu slug)
             'Settings',                                      // Page title
@@ -24,7 +24,7 @@ class Settings
     }
 
     // Callback to render settings page content
-    public function scouting_oidc_settings_page_callback() {
+    public function scouting_oidc_settings_page_callback(): void {
         ?>
         <div class="wrap">
             <h1><?php esc_html_e('Settings', 'scouting-openid-connect'); ?></h1>
@@ -45,7 +45,7 @@ class Settings
     }
 
     // Add the OpenID Connect settings page to the admin menu
-    public function scouting_oidc_settings_page_init() {
+    public function scouting_oidc_settings_page_init(): void {
         $scouting_oidc_settings_oidc = new Settings_Oidc();
         $scouting_oidc_settings_oidc->scouting_oidc_settings_oidc();
 
@@ -56,13 +56,13 @@ class Settings
     /**
      * This script renders JavaScript to hide the custom redirect field when not needed.
      */
-    public function scouting_oidc_fields_enqueue_hide_field_script() {
+    public function scouting_oidc_fields_enqueue_hide_field_script(): void {
         // Enqueue the external JavaScript file
         wp_enqueue_script(
             'hide-field-script',                    // Handle name
             plugins_url('hide-field.js', __FILE__), // Path to the file
             array(),                                // No dependencies
-            "2.2.0",                                // Version number
+            "2.3.0",                                // Version number
             array(
                 'strategy' => 'defer',              // Add the defer attribute
                 'in_footer' => true                 // Load the script in the footer
@@ -71,7 +71,7 @@ class Settings
     } 
 
     // Set up defaults during installation
-    public function scouting_oidc_settings_install() {
+    public function scouting_oidc_settings_install(): void {
         // Set default options for OIDC
         update_option('scouting_oidc_client_id', '');
         update_option('scouting_oidc_client_secret', '');
@@ -85,6 +85,7 @@ class Settings
         update_option('scouting_oidc_user_address', false);
         update_option('scouting_oidc_user_woocommerce_sync', false);
         update_option('scouting_oidc_user_auto_create', true);
+        update_option('scouting_oidc_user_duplicate_email', 'plus_addressing');
         update_option('scouting_oidc_user_redirect', true);
         update_option('scouting_oidc_login_redirect', 'frontpage');
         update_option('scouting_oidc_custom_redirect', '');